+++ examples3d/cross.rs
@@ -3,10 +3,8 @@ extern crate ncollide;
 extern crate nphysics3d;
 extern crate nphysics_testbed3d;
 
-use std::sync::Arc;
 use na::{Pnt3, Vec3, Translation};
-use ncollide::shape::{Plane, Cuboid, Compound};
-use ncollide::inspection::Repr3;
+use ncollide::shape::{Plane, Cuboid, Compound, ShapeHandle};
 use nphysics3d::volumetric::Volumetric;
 use nphysics3d::world::World;
 use nphysics3d::object::RigidBody;
@@ -31,17 +29,17 @@ fn main() {
      */
     let mut cross_geoms = Vec::new();
 
-    let edge_x = Box::new(Cuboid::new(Vec3::new(4.96f32, 0.21, 0.21)));
-    let edge_y = Box::new(Cuboid::new(Vec3::new(0.21f32, 4.96, 0.21)));
-    let edge_z = Box::new(Cuboid::new(Vec3::new(0.21f32, 0.21, 4.96)));
+    let edge_x = Cuboid::new(Vec3::new(4.96f32, 0.21, 0.21));
+    let edge_y = Cuboid::new(Vec3::new(0.21f32, 4.96, 0.21));
+    let edge_z = Cuboid::new(Vec3::new(0.21f32, 0.21, 4.96));
 
-    cross_geoms.push((na::one(), Arc::new(edge_x as Box<Repr3<f32>>)));
-    cross_geoms.push((na::one(), Arc::new(edge_y as Box<Repr3<f32>>)));
-    cross_geoms.push((na::one(), Arc::new(edge_z as Box<Repr3<f32>>)));
+    cross_geoms.push((na::one(), ShapeHandle::new(edge_x)));
+    cross_geoms.push((na::one(), ShapeHandle::new(edge_y)));
+    cross_geoms.push((na::one(), ShapeHandle::new(edge_z)));
 
     let compound = Compound::new(cross_geoms);
     let mass     = compound.mass_properties(1.0);
-    let cross    = Arc::new(Box::new(compound) as Box<Repr3<f32>>);
+    let cross    = ShapeHandle::new(compound);
 
     /*
      * Create the crosses