+++ src/volumetric/volumetric_cone.rs
@@ -23,9 +23,9 @@ pub fn cone_volume<N: Scalar>(dim: usize, half_height: N, radius: N) -> N {
     }
 }
 
-/// The surface of a cone.
+/// The area of a cone.
 #[inline]
-pub fn cone_surface<N: Scalar>(dim: usize, half_height: N, radius: N) -> N {
+pub fn cone_area<N: Scalar>(dim: usize, half_height: N, radius: N) -> N {
     assert!(dim == 2 || dim == 3);
 
     match dim {
@@ -97,8 +97,8 @@ pub fn cone_unit_angular_inertia<N, I>(dim: usize, half_height: N, radius: N) ->
 macro_rules! impl_volumetric_cone(
     ($t: ident, $dim: expr, $p: ident, $i: ident) => (
         impl<N: Scalar> Volumetric<N, $p<N>, $i<N>> for $t<N> {
-            fn surface(&self) -> N {
-                cone_surface($dim, self.half_height(), self.radius())
+            fn area(&self) -> N {
+                cone_area($dim, self.half_height(), self.radius())
             }
 
             fn volume(&self) -> N {