+++ src/volumetric/volumetric_repr.rs
@@ -1,5 +1,5 @@
 use na::{Pnt2, Iso2, Pnt3, Mat3, Iso3, Mat1};
-use ncollide::shape::{Ball, Cone, Cylinder, Convex3, Cuboid2, Cuboid3, Compound2, Compound3};
+use ncollide::shape::{Ball, Cone, Cylinder, Convex2, Convex3, Cuboid2, Cuboid3, Compound2, Compound3};
 use ncollide::inspection::Repr;
 use ncollide::math::Scalar;
 use volumetric::Volumetric;
@@ -39,41 +39,31 @@ macro_rules! dispatch(
 
 impl<N> Volumetric<N, Pnt2<N>, Mat1<N>> for Repr<Pnt2<N>, Iso2<N>>
     where N: Scalar {
-    fn surface(&self) -> N {
-        dispatch!(Pnt2<N>, Mat1<N>, Compound2<N>, Cuboid2<N>/* Convex2<N> */, Cuboid2<N>, self.surface())
-        //                                              XXX:   ^^^^^^^^^^
-        //                                              Change this when Volumetric is implemented for 2D convex.
+    fn area(&self) -> N {
+        dispatch!(Pnt2<N>, Mat1<N>, Compound2<N>, Convex2<N>, Cuboid2<N>, self.area())
     }
 
     fn volume(&self) -> N {
-        dispatch!(Pnt2<N>, Mat1<N>, Compound2<N>, Cuboid2<N>/* Convex2<N> */, Cuboid2<N>, self.volume())
-        //                                              XXX:   ^^^^^^^^^^
-        //                                              Change this when Volumetric is implemented for 2D convex.
+        dispatch!(Pnt2<N>, Mat1<N>, Compound2<N>, Convex2<N>, Cuboid2<N>, self.volume())
     }
 
     fn center_of_mass(&self) -> Pnt2<N> {
-        dispatch!(Pnt2<N>, Mat1<N>, Compound2<N>, Cuboid2<N>/* Convex2<N> */, Cuboid2<N>, self.center_of_mass())
-        //                                              XXX:   ^^^^^^^^^^
-        //                                              Change this when Volumetric is implemented for 2D convex.
+        dispatch!(Pnt2<N>, Mat1<N>, Compound2<N>, Convex2<N>, Cuboid2<N>, self.center_of_mass())
     }
 
     fn unit_angular_inertia(&self) -> Mat1<N> {
-        dispatch!(Pnt2<N>, Mat1<N>, Compound2<N>, Cuboid2<N>/* Convex2<N> */, Cuboid2<N>, self.unit_angular_inertia())
-        //                                              XXX:   ^^^^^^^^^^
-        //                                              Change this when Volumetric is implemented for 2D convex.
+        dispatch!(Pnt2<N>, Mat1<N>, Compound2<N>, Convex2<N>, Cuboid2<N>, self.unit_angular_inertia())
     }
 
     fn mass_properties(&self, density: N) -> (N, Pnt2<N>, Mat1<N>) {
-        dispatch!(Pnt2<N>, Mat1<N>, Compound2<N>, Cuboid2<N>/* Convex2<N> */, Cuboid2<N>, self.mass_properties(density))
-        //                                              XXX:   ^^^^^^^^^^
-        //                                              Change this when Volumetric is implemented for 2D convex.
+        dispatch!(Pnt2<N>, Mat1<N>, Compound2<N>, Convex2<N>, Cuboid2<N>, self.mass_properties(density))
     }
 }
 
 impl<N> Volumetric<N, Pnt3<N>, Mat3<N>> for Repr<Pnt3<N>, Iso3<N>>
     where N: Scalar {
-    fn surface(&self) -> N {
-        dispatch!(Pnt3<N>, Mat3<N>, Compound3<N>, Convex3<N>, Cuboid3<N>, self.surface())
+    fn area(&self) -> N {
+        dispatch!(Pnt3<N>, Mat3<N>, Compound3<N>, Convex3<N>, Cuboid3<N>, self.area())
     }
 
     fn volume(&self) -> N {